move rate limit struct parsing into backend-client
